SOLR-16835: Add approximate-/select to OAS
Now that Solr uses its OAS to generate client bindings in multiple
languages (Java and JavaScript, so far), users of these clients may wish
to use them to run searches.

Normally, this would require converting the `/select` endpoint to JAX-RS
so that its inputs and outputs can be described comprehensively in our
OAS.  However, doing this for `/select` will take some time due to the
complexity and large degree of configurability the endpoint offers.

This commit works around this limitation by creating an approximation of
the `/select` endpoint that can appear in our OAS until the API can be
converted to JAX-RS in earnest.  This will give generated-client users
access to at least some query functionality in this interim.

The `/select` query parameters supported in this commit were chosen
mostly arbitrarily.  They may be added to freely as generated-client
users run into particular needs (e.g. for setting 'faceting'
parameters).