Make 'jsonQuery' interface more flexible
SelectApi.jsonQuery now takes a specially-annotated InputStream rather
than the slightly more structured `Map<String, Object>`.  This
ultimately results in client code that's a little more flexible in how
the JSON query is specified.

With this change, the generated Java client can take either a `String`
or any more structured Jackson-serializable object.  The Javascript
client can take a similar union of structured and unstructured types.

In support of this, this commit introduces a "x-genericEntity" extension
to our OAS, which the Java generation template required.