Test fixes, pt 2
Some S3 tests were continuing to fail after the Jersey upgrade due to
some severe misalignment in the version of jakarta.xml.bind that was
needed by various libraries, with the newer Jersey version relying on
xml.bind 4.0.0, with s3-mock relying on 2.3.3.  This was causing
(test-only) NoClassDefFoundErrors in the s3 tests.

The best solution I could find to this was to stop Jersey from
pulling in xml-bind.  We can always reverse course on this, but for now
Solr uses its own XML serialization/deserialization anyways, so we're
not using xml-bind.

The exclusion itself was pretty straightforward, but required dropping a
few Jersey test dependencies that we weren't making much use of anyways.