Rename withDefaultColl to withDefaultDataStore
This brings the naming for most builders closer into line with the new
SolrRequest.requiresDataStore() method.

Cloud SolrClient builders retain `withDefaultCollection`, both to
preserve backcompat and because it doesn't seem to make sense to make
the builder method more generic in the Cloud case when we know a priori
that cloud users will be interacting with collections primarily.