SOLR-17116: Fix INSTALLSHARDDATA async reporting
Prior to this commit, failures in asynchronous 'INSTALLSHARDDATA'
commands weren't properly detected.  Instead the task was always labeled
as 'completed', even when it failed.

Ultimately, this occurred because the overseer processing for this
command ignored the output of per-core requests.  Incorporating these
responses into the overseer "results" allows task operation to be
correctly classified and reported.