SOLR-17066: Replace 'data store' term in code and docs (#2201)
"Data store" as a term, was introduced a few weeks back as a way to
refer generically to something that could be either a collection or a
core.  As its use expanded though, it turned out to not be very well
liked in practice.

This commit replaces all usage of the term with either "collection" or
"index type", depending on the specific usage.