SOLR-17137 - Add configuration for SSL between Solr and Prometheus exporter.
When running Solr with mTLS the exporter also needs to use mTLS to be
able to talk to Solr api.

I have used the same env-var/jvm properties as expected in Solr. This is
simply because we have Solr and Prometheus exporter running in same Pod
in k8s.

Basic auth credentials are added via env variable in
solr-exporter{,.cmd} starter scripts and documented in the reference
guide.

*Added a simple bats test for Solr prometheus exporter*

I added the bin-folder for prometheus-exporter into PATH to be able
to call the startup script. I also added a port env var for the exporter
based on the Solr port utilizing the same convention as zookeeper port.

Since the prometheus exporter is not a part of solr cli I had to
find way to stop the process. The simplest way I found was to grep for
the process and kill the pid. I put this in a helper function so that
the bats test is clean.

The test verifies that solr exporter actually scraps the running
solr instance by verifying information about a newly make core is present.