SOLR-17066: Add GenericSolrRequest.setRequiresCollection (#2229)
SOLR-17066 added a 'defaultCollection' field to each SolrClient
implementation, similar to the one that has long been in use for SolrJ's
"cloud" clients.  This default collection (or core) is used on a
request-by-request basis to build the complete HTTP path, based on the
value of SolrRequest.requiresCollection().

This is a particular challenge for GenericSolrRequest though, which can
be used to make both collection-agnostic and collection-aware requests.
This commit adds a GenericSolrRequest setter,
`setRequiresCollection(boolean)`, to allow GSR users to specify whether
the client-level default collection should be used on their request or not.