SOLR-17066: Switch HSClient away from core URLs, pt 3
SOLR-17066 introduced a `withDefaultCollection` builder method that can
be used in lieu of providing a core URL as a "base".  (Core URLs are not
ideal as they prevent clients from making core-agnostic requests.)

However, much of Solr's own client usage still needs modified to use
this new builder method.

This commit switches test code in the 'solrj' module over to using
`withDefaultCollection`.

Check and tests both pass as of this commit.