Reinstate "SOLR-17066: Switch HttpSolrClient away from coreURLs, pt 2 (#2231)""
This commit fixes a bug in a previously committed (and then reverted)
change e8ac06.

The root of the problem lies in how Solr handles context-roots.  Solr 9
allows configurable and arbitrary context roots, while Solr 10 disallows
this practice.

This difference wasn't accounted for when backporting the core URL
removal to branch_9x, causing a large number of test failures due to
some assumptions around what a path should look like.  These assumptions
were valid for Solr 10 (i.e. `main`) but invalid for Solr 9.

In order to reinstate this commit on branch_9x, the commit no longer
inspects and splits user-provided URLs when creating clients.  Instead,
it relies on thorough javadocs to clarify that use of core URLs is
(1) deprecated, and (2) mutually exclusive with the
`withDefaultCollection` builder option.