SOLR-17068: Add missing PostToolTest @SuppressSSL
The test on 'main' was initially added with a SuppressSSL annotation,
but it appears to have been lost at some point in the backport process.
As a result, PostToolTest fails ~6% of the time on 9x with SSL related
issues.

We should better understand the cause of this SSL failure, as we should
for all usages of @SuppressSSL, but for now this commit re-adds the
annotation that was errantly dropped.