Fix LBHttp2SolrClient zombie testing
Zombie-resurrection was broken in LBHttp2SolrClient because the default
collection wasn't being taken into account on these requests.  The root
of the problem is the way that LBHttp2 takes in an already-constructed
SolrClient, making it difficult to set the default collection.  (In
contrast, LBHttpSolrClient creates a different client for each base URL
using a common Builder - this allows it to property set the default
collection.)

This commit fixes this by created a 'DelegatingSolrClient' that LBHttp2
can use internally around its single Http2SolrClient.  The default
collection is set on this 'Delegating' wrapper, so that it gets picked
up by the zombie-resurrection logic.