WIP - refactor LB code to rely on endpoints
This acts out the first step of the 3 below that I've been considering
to make the LB changes a little more reasonably.

I _haven't_ actually switched any LB usage over to using base URLs or
anything either.  Just trying to match the existing behavior and usage,
with this 'Endpoint' abstraction jammed in.

1. Create 'Endpoint' class (in LBSolrCLient, or is this generally more
   useful?)
2. Change LBSolrClient.ServerWrapper to be an "EndpointWrapper"
3. Change internal LBSolrClient logic to use endpoint instead of Server