SOLR-17066: Add deprecation warnings to 9x methods
SOLR-17066 introduced the `LBSolrClient.Endpoint` class, and modified
several "LB" client methods to use it on "main".  This commit adds the
corresponding deprecation notices to branch_9x, so that users can be
notified of these changes and start to upgrade their apps where
appropriate/required.

This commit should only be added to branch_9x.