update comment in test
In earlier manual testing of subqueries with basic auth enabled, as part of SOLR-12813, errors were seen when forwardCrendentials was set to false; the comment here reflected this. Likely those errors were a result of temporary code changes that were not committed. In subsequent manual testing, those errors have not recurred. This test passes regardless of whether forwardCrentials is true or false, and recent manual testing has had the same outcome.