include user Principal when building SolrQueryRequest
EmbeddedSolrServer#request() has two separate codepaths where a SolrQueryRequest is created using the _parser.buildRequestFrom() utility method. The first codepath is active when the relevant SolrRequestHandler can be gotten from the CoreContainer. The second codepath is active when coreContainer.getRequestHandler(path) returns null and instead we have to get the SolrRequestHandler directly from the SolrCore. This second codepath is the one that's used in subquery execution. It was updated in the initial fix for SOLR-12813 so that the call to _parser.buildRequestFrom() would now include the user Principal. However, the first codepath was left alone because it was not found to be involved in subquery execution. In the present commit, the first codepath is being updated as well. This change is not needed for addressing the issue described in SOLR-12813, but it is being made in the interest of keeping the logic as consistent as possible across the two codepaths in EmbeddedSolrServer.request()