SOLR-17192: Add "field-limiting" URP to catch ill-designed schemas (#2395)
The URPF, `NumFieldLimitingUpdateRequestProcessorFactory`, blocks all
update requests that go through `processAdd` if the core exceeds a
configurable threshold of fields.

The factory accepts two parameters: `maxFields` is a required integer
representing the maximum field threshold,  and `warnOnly` is an optional
boolean that (when enabled) has the URP chain log warnings instead of
blocking updates.

The factory is included in the default configset, with warnOnly=true and
maxFields=1000.  ('warnOnly' will be false starting in the 10.0 release)

---------

Co-authored-by: David Smiley <dsmiley@apache.org>