SOLR-17195: Add 'minPrefixLength' soft limit
Prefix queries have a memory cost that increases in proportion with the
number of indexed terms that start with that prefix.  This can cause
stability problems, especially for short prefixes which can match a
large number of indexed-terms.

This commit introduces a new solrconfig.xml setting, `minPrefixLength`,
similar to the existing `maxBooleanClauses`.  This setting causes an
error to be thrown any time a prefix query is created with a prefix
whose length is less than the configured minumum.

This setting is set at '2' by default, disallowing only single-character
prefix queries.

Still TODO:
  - ref guide documentation
  - more error-case testing
  - handle the faceting invocation
  - consider whether this should be disable-able on a per request basis