Convert filestore 'get' API to JAX-RS
Plausibly works on the server-side, though the generated SolrRequest
implementation will almost surely NOT work.

Finally, this commit changes the path for these APIs to be
/cluster/files (instead of the previously used `/node/files`).  I need
to think through the backcompat implications of this a bit more
thoroughly.  I probably need a version of these APIs also available at
the original path.