SOLR-17195: Add 'minPrefixLength' soft limit (#2499)
Prefix queries have a memory cost that increases in proportion with the
number of indexed terms that start with that prefix.  This can cause
stability problems, especially for short prefixes which can match a
large number of indexed-terms.

This commit introduces a new solrconfig.xml setting,
`minPrefixQueryTermLength`, similar to the existing
`maxBooleanClauses`.  This setting causes an error to be thrown any
time a prefix query is created with a prefix shorter than the configured
minumum.

The limit is set at '-1' in the default configset, essentially disabling it.  But
users may change the value in their own configs or use the
'solr.query.minPrefixLength' property as an override.  The limit may also be
overridden on a per-query basis by providing a 'minPrefixQueryTermLength'
local param.