Default to MockDirectoryFactory in test configs (#2598)
This was already the effective default value, by virtue of it being set
in 'randomization.gradle'.  But setting the value there instead of in
the config files themselves meant that the value wasn't being properly
set for IDE test runners.

This commit fixes this by removing the 'randomization.gradle' setting
and instead setting MockDirectoryFactory as the default (via ${}
expansion syntax) in each individual test solrconfig file.