Review SolrJ code for typos (#2656)
Fixing typos and grammer errors.  These are mostly in the various comments and javadocs.  There are a few methods that are mispelled, fixed the private ones, and added deprecation for the non private ones.  Have not changed variable names that are flagged as typos by Intellij.

---------

Co-authored-by: Jan Høydahl <janhoy@apache.org>
(cherry picked from commit 9589cbc2338cb0a526363d33536223424578dbab)