Remove invalid usage of "scheme" option in bin/solr config
IMPORTANT: Users have to provide the Solr URL scheme via the Solr URL parameter or configure it via the environment variable solr.url.scheme in case where they use the default Solr URL.

Note: This breaking change was introduced accidentally in 9.7. This commit fixes the errors thrown by the breaking change without introducing backwards compatibility with 9.6. This is due to the overlapping option flags (-s) and the already applied changes in the CLI files.