SOLR-17413: ulog replay should copy SolrQueryRequest
SolrQueryRequest is a non-threadsafe type, but was being shared across
executor threads during UpdateLog replay.  This introduces a number of
issues, not the least being a ConcurrentModificationException if
multiple threads happen to tweak the request 'context' simultaneously.

This commit fixes this issue by giving each executor thread a unique
LocalSolrQueryRequest instance to use.