SOLR-16825: Migrate v2 definitions to 'api' module, pt 8
SOLR-16825 added a new gradle module, 'api', which holds v2 API
definitions as interfaces.  This allwos us to generate an OAS (and other
artifacts) as a part of the solrj build.  But these artifacts only cover
the v2 APIs present in the 'api' module.

This commit moves v2 API defining annotations to new interfaces in the
'api' module for several "schema" fetching and listing APIs.