SOLR-17456: TransactionLog ctor integrity (#2762)
The TransactionLog constructor can't handle an existing file being present; it shouldn't be there.
Should throw an exception in this case, NOT log a warning which would leave the object in a partially constructed state.
This should happen in the first place, of course.  I see no evidence it has occurred.