Refactor collection-snapshot code to isolate resp
Prior to this CollectionSnapshotMetaData and CoreSnapshotMetadata
contained application/serialization logic, and references to other
'core' classes, making it hard for our OpenAPI spec to model these
responses.

This commit attempts to refactor the snapshot code so that the response
objects live in our 'api' module, with the related logic remaining in
core.

Has not been tested or validated yet.