Still trying to get TestSolrCloudSnapshots working
My guess is that whatever is writing the data into ZK is not properly
getting the enum value, but is instead writing out a sketchy toString
value that then fails to deserialize when read later on.

From some unit testing it looks like Jackson can serialize the enum just
fine. I'm suspecting it's Javabin that can't, which is the default
format used for SolrJ requests.  I could probably fix this by having the
test use the generated SolrRequest, but that would leave things broken
for anyone not using that class, so it's not a real fix.

I'm gonna put a pin in things for now, but I can probably return to this
later on if I get time.