Fix ZookeeperReadAPITest Failure
Without the '@Produces' annotation on the interface, the API was trying
to respond with 'javabin', which caused the test to fail.  This commit
adds back in the '@Produces' annotations, so that Solr knows to use
RawResponseWriter to write out the file content.

I don't think we want to use '@Produces' generally, but this API is a
reasonable exception.