SOLR-16781: Remove solrconfig.xml <lib> directives (#2875)
Solr offers a number of ways for users to add JARs and resources to
their classpath, including:

* solr.xml <sharedLib> entries
* SOLR_MODULES env-var/support
* core and install-level "lib/" directories
* the package manager
* direct classpath modification

In addition to being largely redundant with the methods above,
solrconfig.xml's <lib> directive has been a pain point and source of
security concerns in the past.  This commit removes it from Solr 10.