SOLR-16391: Convert "modify-coll" API to JAX-RS
In addition to JAX-RS conversion (which adds the API to our v2 OAS, and
generated clients), this PR also makes a few modifications to the v2
form to make the API fit our more REST-ful paradigm.  In particular, the
API has changed to a 'PUT', and the "modify" command property and
additional nesting in the request body have been removed.

Still needed:
  - check and tests
  - ref-guide changes
  - CHANGES.txt
  - discussion of "unset" semantics and syntax.