Create API for raw file fetching
Identical to the pre-existing`/node/files/a/b.txt`, but in its own
endpoint.  Does NOT remove the existing endpoint or switch tests over
to use it.  Those can be handled in a separate commit to enable this one
to be easily backported later to 9x if desired.

This also doesn't enable codegen for the new API, though I think the
blockers are sufficiently cleared up now to enable that if desired.