Temporary stash to investigate param screening behavior
TestDistribFileStore fails with the changes in this commit.  The changes
themselves are fine, but it fails because of a pre-existing *ahem*
feature in SolrClient where SolrParams are encoded as form-params
instead of query-params in some circumstances.

I could probably hack around this for the moment (e.g. by using a
builder setter to ensure 'getFrom' is sent as a true query param).

But I've decided to halt progress on this branch and break some other
pieces out into their own PRs, while I send some questions out to the
community on this front.