SOLR-17677: Ensure DBQ is safe before running
Portions of the DBQ codepath execute the deletion query with a standard
Lucene IndexSearcher, which upsets some Solr Query implementations that
have been built to assume SolrIndexSearcher in (e.g.)
Query.createWeight.

This commit adds checks for this case, which work by using a Lucene
"QueryVisitor" to iterate over a parsed query tree and detect whether
any of the individual queries require SolrIndexSearcher.