Fix small race condition in DeleteReplicaTest
The CoreContainer will be provided to the jetty after the jetty is
running. If the test tries to get the cores right after the server is
ready, it will cause a NullPointerException. So just wait on the cores
being ready too.

(cherry picked from commit 8160f9ffff86ab34fb92c6f8fbd1de200060a5f3)