Remove dead code in JavaBinUpdateRequestCodec (#3207)
Dead code removal in JavaBinUpdateRequestCodec.
Compatibility: It's probable that an old SolrJ might sometimes send data encoded in a way that's no longer understood. It's also possible a user is using JavaBinCodec directly to do the same. Henceforth, they will get an error. As we have very little backwards compatibility tests, I can't know for sure.  Regardless, in a major version (Solr 10) it's reasonable to take such an action.