allow rate-limiting on data nodes (for shards.tolerant=true) (#239)
* allow rate-limiting on data nodes (for shards.tolerant=true)

this is a quick-and-dirty hack that will work for our
usage, but this should be reconsidered and something
more general committed upstream.

Honestly it doesn't make sense to hardcode handling of
this at the level of RateLimitManager. Individual rate
limiters should be specified as plugins, with the
context necessary to make their own decisions.

* support sysprop-based toggling of ratelimiting behavior wrt shards.tolerant

adds `solr.dataNodeAllowRateLimitAlways` param (true/false,
default false. If false, data node rateLimiting is only
applied when `shards.tolerant=true`, otherwise (if `true`),
rateLimiting may always be applied on data node for distrib
requests, even if `shards.tolerant=false`.

* move logic into header propagation SearchHandler

* upgrade github workflow `actions/cache` v2->v4

https://github.blog/changelog/2024-12-05-notice-of-upcoming-releases-and-breaking-changes-for-github-actions/#actions-cache-v1-v2-and-actions-toolkit-cache-package-closing-down

* add Override annotation

* fix bug in shards.tolerant parsing