SOLR-17043: Remove SolrClient path pattern matching (#3238)
Prior to this commit, the LB and Cloud clients looked at the URL path to make
some decisions around request routing, etc.  This logic was brittle and not 
really scalable to the more varied paths in Solr's v2 APIs.  This commit
replaces that logic to instead use a "request type" enum already that
SolrRequest objects make available.

---------

Co-authored-by: jmuriithi3 <jmuriithi3@bloomberg.net>
Co-authored-by: Jason Gerlowski <gerlowskija@apache.org>