SOLR-17758: Fix limiting-URP "warnOnly" mode
Due to a bug in the initial implementation, the "warnOnly" mode of the
NumFieldLimiting "URP" never worked as designed.  The "warn" codepath
failed to call `propagateAdd` for the next URP in the chain.  This makes
the "warn only" behavior indistinguishable from the "hard limit"
behavior from a user perspective.

This commit fixes this bug.  It also adds integration testing for
several "warn only" mode related scenarios, to prevent against future
regressions.