SOLR-17661: Fix flaky TestLTRQParserPlugin tests
The flaky tests are:
- TestLTRQParserPlugin.ltr_expensiveFeatureRescoring_shouldTimeOutAndReturnPartialResults
- TestLTRQParserPlugin.ltr_expensiveFeatureRescoringAndPartialResultsNotTolerated_shouldRaiseException

These tests have to timeout after the initial search is complete, and
during the reranking phase. The reranking model takes 1 second to load,
so increase the timeout to 800ms so that the search is more likely to
complete in that timeframe.

(cherry picked from commit 2e159066ee43294cd6b9298df85e9b2901377b10)