SAI-5698: make SolrIndexSearcher.getAndCacheDocSet public, and allow to accept arbitrary cache arg (#271)
SolrIndexSearcher.getAndCacheDocSet() currently operates directly against filterCache. It is trivial to adjust this method to have the cache passed in as an arg. This allows consolidation of logic around QueryLimits.

This could be upstreamed, but if there's objection upstream and/or this becomes a maintenance burden it will be trivial to revert and simply copy the relevant functionality.