SOLR-17896: Introduce reusable navigation tabs (#3656)
Implements the navigation tabs component  in a reusable way and applies it to Configsets and Cluster for demonstration.

The implementation does not apply styling from the designs, because the used composables from Material are not easily customized. The current implementation is also more accessible, as the expected designs used only a color indicator for the current selection, and the implementation uses an indicator line.

This PR is based on #3521.

---------

Co-authored-by: Sanjay Dutt <sanjaydutt@apache.org>