SOLR-17936: add focus indicator to navigation sidebar entries
Rather than using Tab, which does not have colors for our various states,
we use Button, which has all colors except for the Selected state.
Material 3's new "Expressive" version also provides an experiment ToggleButton
component, but our version of the Material 3 dependency is not new enough to
be able to use it (and if I understand correctly, we can't use the very newest
material 3 release, since Compose Multiplatform does not yet support it).

With thanks to @malliaridis for the guidance getting started, and for
suggesting that we try a component other than Tab.