SOLR-17972: Retry creation of ZK lock on connection loss. (#3968)
This makes sure we don't skip creation of ZK distributed lock in case of a transient connection loss. This fix is only for when Solr is running with no overseer (distributed updates).
This change is for 9x branch only, as this parameter is now ignored in 10 with the move to Curator.