Migrate Solr tests from SolrJettyTestBase to using SolrJettyTestRule (#3947)
Embrace the use of @ClassRule setup to manage the Solr lifecycle using SolrClientTestRule in our tests.  This PR is to finish the migration to using SolrJettyTestRule, which extends SolrClientTestRule across our tests.

There is also plenty of small cleanups of the tests throughout.  Lint cleanups and places where we had extra setup steps that did not need to happen for the goal of a individual test to be accomplished.